+++ pioreactor/types.py
@@ -100,6 +100,7 @@ PdAngleOrREF = t.Union[PdAngle, t.Literal["REF"]]
 AnalogValue = t.Union[int, float]
 Voltage = float  # maybe should be non-negative?
 OD = t.Annotated[float, Meta(ge=0)]
+OD600 = t.Annotated[float, Meta(ge=0)]
 
 AdcChannel = t.Literal[0, 1, 2, 3]